stackcollapse-git-tr2-event.pl: add an analyzer for git "trace2" events
Git version 2.22.0, which is about to be released in the next few
weeks, has the ability to emit trace logging as JSON.

So this adds the ability to generate a flamegraph of where the time is
spent. As noted in various TODOs there's a lot more that can be done
here, but this already works well enough to use it.